refactor: Move graph creation to dbt_project instead of as a method on ResourceGrouper
